La til eksempelfiler som ble presentert i oevingstime
